+++ packages/falcon-server-env/src/cache/Cache.ts
@@ -26,11 +26,15 @@ export type GetCacheOptions = {
   options?: SetCacheOptions;
 };
 
+const DEFAULT_TAG_TTL: number = 60 * 60; // 1 hour
+
 /**
  * Cache-wrapper with extended methods
  */
 export default class Cache<V = any> implements KeyValueCache<V> {
-  constructor(private cacheProvider: KeyValueCache<V>) {}
+  protected activeGetRequests: Map<string, Promise<V>> = new Map();
+
+  constructor(public provider: KeyValueCache<string>, protected tagTtl: number = DEFAULT_TAG_TTL) {}
 
   async get(key: string): Promise<V>;
 
@@ -38,45 +42,20 @@ export default class Cache<V = any> implements KeyValueCache<V> {
 
   /**
    * Returns cached value for the provided key and setOptions object
-   * @param {string} key Cache key
-   * @param {GetCacheOptions} setOptions Object with params to fetch the data to be cached
-   * @returns {Promise<string|undefined>} Cached value
+   * @param key Cache key
+   * @param setOptions Object with params to fetch the data to be cached
+   * @returns Cached value
    */
   async get(key: string, setOptions?: GetCacheOptions): Promise<V> {
-    let value: GetCacheFetchResult = await this.cacheProvider.get(key);
-    // Validating by cache tags
-    if (this.isValueWithOptions(value)) {
-      const { tags: tagMap = {} } = value.options as ValueOptions;
-      if (await this.isTagMapValid(tagMap as TagMap)) {
-        ({ value } = value);
-      } else {
-        // If tags are invalid - set value as "not found"
-        value = undefined;
-        await this.delete(key);
-      }
+    if (this.activeGetRequests.has(key)) {
+      return this.activeGetRequests.get(key) as Promise<V>;
     }
 
-    if (typeof value === 'undefined' && typeof setOptions === 'object') {
-      const { fetchData } = setOptions;
-      let { options } = setOptions;
-      const cacheResult: GetCacheFetchResult = await fetchData();
-      if (typeof cacheResult !== 'undefined') {
-        if (this.isValueWithOptions(cacheResult)) {
-          ({ value } = cacheResult);
-          const { options: cacheResultOptions = {} } = cacheResult;
-          // Merging cache options from the "fetchData" result and passed method argument
-          options = Object.assign({}, options, cacheResultOptions);
-        } else {
-          value = cacheResult;
-        }
+    this.activeGetRequests.set(key, this.createGetRequest(key, setOptions));
+    const result = await this.activeGetRequests.get(key);
+    this.activeGetRequests.delete(key);
 
-        if (typeof value !== 'undefined') {
-          await this.set(key, value, options as SetCacheOptions);
-        }
-      }
-    }
-
-    return value;
+    return result as V;
   }
 
   async set(key: string, value: V): Promise<void>;
@@ -95,7 +74,13 @@ export default class Cache<V = any> implements KeyValueCache<V> {
         }
       };
     }
-    return this.cacheProvider.set(key, cachedValue, options);
+
+    if (Array.isArray(cachedValue) || typeof cachedValue === 'object') {
+      // For non-scalar values - JSON.stringify
+      cachedValue = JSON.stringify(cachedValue);
+    }
+
+    return this.provider.set(key, cachedValue, options);
   }
 
   async delete(key: string): Promise<boolean>;
@@ -113,7 +98,7 @@ export default class Cache<V = any> implements KeyValueCache<V> {
       await Promise.all(key.map(kKey => this.delete(kKey)));
       return;
     }
-    return this.cacheProvider.delete(key);
+    return this.provider.delete(key);
   }
 
   /**
@@ -152,7 +137,7 @@ export default class Cache<V = any> implements KeyValueCache<V> {
         if (typeof tagValue === 'undefined' && createIfMissing) {
           // For "createIfMissing" flag - generate new tag value and save it to the cache
           tagValue = this.generateTagValue();
-          await this.set(tag, tagValue);
+          await this.set(tag, tagValue, { ttl: this.tagTtl });
         }
 
         if (tagValue) {
@@ -173,14 +158,59 @@ export default class Cache<V = any> implements KeyValueCache<V> {
     return typeof data === 'object' && 'value' in data && 'options' in data;
   }
 
+  private async createGetRequest(key: string, setOptions?: GetCacheOptions): Promise<V> {
+    let value: GetCacheFetchResult = await this.provider.get(key);
+    try {
+      value = JSON.parse(value);
+    } catch {
+      // Keep `value` with the original value
+    }
+
+    // Validating by cache tags
+    if (this.isValueWithOptions(value)) {
+      const { tags: tagMap = {} } = value.options as ValueOptions;
+      if (await this.isTagMapValid(tagMap as TagMap)) {
+        ({ value } = value);
+      } else {
+        // If tags are invalid - set value as "not found"
+        value = undefined;
+        await this.delete(key);
+      }
+    }
+
+    if (typeof value === 'undefined' && typeof setOptions === 'object') {
+      const { fetchData } = setOptions;
+      let { options } = setOptions;
+      const cacheResult: GetCacheFetchResult = await fetchData();
+      if (typeof cacheResult !== 'undefined') {
+        if (this.isValueWithOptions(cacheResult)) {
+          ({ value } = cacheResult);
+          const { options: cacheResultOptions = {} } = cacheResult;
+          // Merging cache options from the "fetchData" result and passed method argument
+          options = Object.assign({}, options, cacheResultOptions);
+        } else {
+          value = cacheResult;
+        }
+
+        if (typeof value !== 'undefined') {
+          await this.set(key, value, options as SetCacheOptions);
+        }
+      }
+    }
+
+    return value;
+  }
+
   /**
    * Generating a short and safe enough tag value (second + ms = will ensure a unique value for the same tag name)
-   * @returns {string} Tag value
+   * @returns {number} Tag value
    */
-  private generateTagValue(): string {
-    const date: Date = new Date();
-    return `${Date.now()
-      .toString()
-      .substr(6)}`;
+  private generateTagValue(): number {
+    return parseInt(
+      Date.now()
+        .toString()
+        .substr(6),
+      10
+    );
   }
 }