+++ packages/falcon-server/src/schemaDirectives/GraphQLCacheInvalidatorDirective.js
@@ -35,17 +35,10 @@ module.exports = class GraphQLCacheInvalidatorDirective extends SchemaDirectiveV
    * @returns {void}
    */
   visitFieldDefinition(field) {
-    let { resolve = defaultFieldResolver } = field;
+    const { resolve = defaultFieldResolver } = field;
     const { idPath } = this.args;
 
-    Object.defineProperty(field, 'resolve', {
-      get: () => this.getResolverWithCacheInvalidator(resolve, idPath),
-      // Handling potential "addResolveFunctionsToSchema" calls that define dynamic resolvers
-      set: newResolve => {
-        resolve = newResolve;
-      },
-      configurable: true
-    });
+    field.resolve = this.getResolverWithCacheInvalidator(resolve, idPath);
   }
 
   /**
@@ -79,8 +72,8 @@ module.exports = class GraphQLCacheInvalidatorDirective extends SchemaDirectiveV
    * Invalidate cache from the result using the provided idPath entry
    * @param {mixed} result Resolver result
    * @param {IdPathEntry} idPathEntry ID Path entry
-   * @param {Object} parent GraphQL Resolver parent value
-   * @param {Object} context GraphQL context object
+   * @param {object} parent GraphQL Resolver parent value
+   * @param {object} context GraphQL context object
    * @param {GraphQLResolveInfo} info GraphQL Info object
    * @returns {void}
    */